+++ back/__tests__/integration/database/queries/user.queries.test.js
@@ -9,15 +9,12 @@ const user = {
     hashed_password: "password",
 };
 
-beforeEach(async () => {
+beforeAll(async () => {
     await prisma.users.create({ data: user });
 });
 
-afterEach(async () => {
-    await prisma.users.delete({ where: user });
-});
-
 afterAll(async () => {
+    await prisma.users.delete({ where: { userID: user.userID } });
     await prisma.$disconnect();
 });
 
@@ -64,6 +61,60 @@ describe("Get user by id", () => {
 
     it("should return null if user does not exist", async () => {
         const response = await userQueries.getUserById(0);
+        expect(response).toBeNull();
+    });
+});
+
+describe("Get user password by id", () => {
+    it("should return user password if user exists", async () => {
+        const response = await userQueries.getUserPasswordById(user.userID);
+        expect(response).toBe(user.hashed_password);
+    });
+
+    it("should return null if user does not exist", async () => {
+        const response = await userQueries.getUserPasswordById(0);
+        expect(response).toBeNull();
+    });
+});
+
+describe("Get user type by id", () => {
+    it("should return Admin if user is an admin", async () => {
+        await prisma.admins.create({ data: { user: { connect: { userID: user.userID } } } });
+        const response = await userQueries.getUserTypeById(user.userID);
+        expect(response).toBe("Admin");
+        await prisma.admins.delete({ where: { adminID: user.userID } });
+    });
+
+    it("should return Teacher if user is a teacher", async () => {
+        await prisma.teachers.create({ data: { teacher_number: -1, user: { connect: { userID: user.userID } } } });
+        const response = await userQueries.getUserTypeById(user.userID);
+        expect(response).toBe("Teacher");
+        await prisma.teachers.delete({ where: { teacherID: user.userID } });
+    });
+
+    it("should return Student if user is a student", async () => {
+        await prisma.students.create({ data: { student_number: -1, user: { connect: { userID: user.userID } } } });
+        const response = await userQueries.getUserTypeById(user.userID);
+        expect(response).toBe("Student");
+        await prisma.students.delete({ where: { studentID: user.userID } });
+    });
+
+    it("should return null if user is in multiple roles", async () => {
+        await prisma.admins.create({ data: { user: { connect: { userID: user.userID } } } });
+        await prisma.teachers.create({ data: { teacher_number: -1, user: { connect: { userID: user.userID } } } });
+        const response = await userQueries.getUserTypeById(user.userID);
+        expect(response).toBeNull();
+        await prisma.admins.delete({ where: { adminID: user.userID } });
+        await prisma.teachers.delete({ where: { teacherID: user.userID } });
+    });
+
+    it("should return null if user is not an admin, teacher or student", async () => {
+        const response = await userQueries.getUserTypeById(user.userID);
+        expect(response).toBeNull();
+    });
+
+    it("should return null if user does not exist", async () => {
+        const response = await userQueries.getUserTypeById(0);
         expect(response).toBeNull();
     });
 });
